+++ app/scripts/OSMTilesTrack.js
@@ -22,8 +22,7 @@ class OSMTilesTrack extends PixiTrack {
    */
   constructor(scene, options, animate) {
     // Force OpenStreetMaps copyright
-    options.name = `© OpenStreetMap${options.name ? `\n${options.name}` : ''}`;
-
+    // options.name = `© OpenStreetMap${options.name ? `\n${options.name}` : ''}`;
     super(scene, options);
 
     // the tiles which should be visible (although they're not necessarily fetched)