+++ app/scripts/HorizontalGeneAnnotationsTrack.js
@@ -78,6 +78,9 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
 
     tile.tileData.forEach((td, i) => {
       const geneInfo = td.fields;
+      const geneName = geneInfo[3];
+      const geneId = this.geneId(geneInfo);
+
       let fill = this.options.plusStrandColor || 'blue';
 
       if (geneInfo[5] === '-') {
@@ -89,9 +92,8 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
       // don't draw texts for the latter entries in the tile
       if (i >= MAX_TEXTS) return;
 
-      // geneInfo[3] is the gene symbol
       const text = new PIXI.Text(
-        geneInfo[3],
+        geneName,
         {
           fontSize: `${this.fontSize}px`,
           fontFamily: FONT_FAMILY,
@@ -105,7 +107,7 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
       text.anchor.x = 0.5;
       text.anchor.y = 1;
 
-      tile.texts[geneInfo[3]] = text; // index by geneName
+      tile.texts[geneId] = text; // index by geneName
 
       tile.textGraphics.addChild(text);
     });
@@ -174,6 +176,10 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
 
   drawTile() {}
 
+  geneId(geneInfo) {
+    return `${geneInfo[0]}_${geneInfo[1]}_${geneInfo[2]}_${geneInfo[3]}`;
+  }
+
   renderTile(tile) {
     if (!tile.initialized) return;
 
@@ -229,6 +235,9 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
 
         let yMiddle = this.dimensions[1] / 2;
         const geneName = geneInfo[3];
+        const geneId = this.geneId(geneInfo);
+
+        // console.log('geneInfo:', geneInfo);
 
         if (geneInfo[5] === '+') {
           // genes on the + strand drawn above and in a user-specified color or the
@@ -302,7 +311,7 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
         // don't draw texts for the latter entries in the tile
         if (i >= MAX_TEXTS) return;
 
-        const text = tile.texts[geneName];
+        const text = tile.texts[geneId];
 
         if (!text) return;
 
@@ -312,13 +321,13 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
           fill: fill[geneInfo[5]]
         };
 
-        if (!(geneInfo[3] in tile.textWidths)) {
+        if (!(geneId in tile.textWidths)) {
           text.updateTransform();
           const textWidth = text.getBounds().width;
           const textHeight = text.getBounds().height;
 
-          tile.textWidths[geneInfo[3]] = textWidth;
-          tile.textHeights[geneInfo[3]] = textHeight;
+          tile.textWidths[geneId] = textWidth;
+          tile.textHeights[geneId] = textHeight;
         }
       });
   }
@@ -478,7 +487,11 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
 
           const geneInfo = td.fields;
           const geneName = geneInfo[3];
-          const text = tile.texts[geneName];
+          const geneId = this.geneId(geneInfo);
+
+          const text = tile.texts[geneId];
+
+          // console.log('visible geneName:', geneName);
 
           if (!text) return;
 
@@ -507,10 +520,25 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
           text.position.x = this._xScale(txMiddle);
           text.position.y = textYMiddle;
 
+          if (!tile.textWidths[geneId]) {
+            // if we haven't measured the text's width in renderTile, do it now
+            // this can occur if the same gene is in more than one tile, so its
+            // dimensions are measured for the first tile and not for the second
+            const textWidth = text.getBounds().width;
+            const textHeight = text.getBounds().height;
+       
+            // console.log('textWidth:', textWidth);
+            // console.log('textHeight:', textHeight);
+
+            tile.textHeights[geneId] = textHeight;
+            tile.textWidths[geneId] = textWidth;
+          }
+
           if (!parentInFetched) {
             text.visible = true;
 
             const TEXT_MARGIN = 3;
+
 
             if (this.flipText) {
               // when flipText is set, that means that the track is being displayed
@@ -518,15 +546,17 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
               this.allBoxes.push([
                 text.position.x,
                 textYMiddle - fontSizeHalf - 1,
-                text.position.x + tile.textHeights[geneInfo[3]] + TEXT_MARGIN,
-                textYMiddle + fontSizeHalf - 1
+                text.position.x + tile.textHeights[geneId] + TEXT_MARGIN,
+                textYMiddle + fontSizeHalf - 1,
+                geneName
               ]);
             } else {
               this.allBoxes.push([
                 text.position.x,
                 textYMiddle - fontSizeHalf - 1,
-                text.position.x + tile.textWidths[geneInfo[3]] + TEXT_MARGIN,
-                textYMiddle + fontSizeHalf - 1
+                text.position.x + tile.textWidths[geneId] + TEXT_MARGIN,
+                textYMiddle + fontSizeHalf - 1,
+                geneName
               ]);
             }
 
@@ -536,6 +566,7 @@ class HorizontalGeneAnnotationsTrack extends HorizontalTiled1DPixiTrack {
               caption: geneName,
               strand: geneInfo[5]
             });
+
             allTiles.push(tile.textBgGraphics);
           } else {
             text.visible = false;